engine: extract esm/script_interface.js
