+++ engine/esm/transpiled.js
@@ -2,13 +2,11 @@ import { ss } from "./ss.js";
 import { registerType } from "./typesystem.js";
 
 import {
-  Util,
   Mouse,
 } from "./util.js";
 
 import {
   globalRenderContext,
-  useGlVersion2,
   set_globalRenderContext,
   set_tilePrepDevice,
   set_useGl,
@@ -18,10 +16,11 @@ import {
 import {
   globalScriptInterface,
   globalWWTControl,
-  layerManagerGetAllMaps,
+  loadWtmlFile,
   set_freestandingMode,
   set_globalScriptInterface,
   set_globalWWTControl,
+  set_loadWtmlFile,
 } from "./data_globals.js";
 
 import { BlendState } from "./blend_state.js";
@@ -41,12 +40,9 @@ import { Sprite2d } from "./graphics/sprite2d.js";
 
 import { Coordinates } from "./coordinates.js";
 
-import { Annotation, Circle, Poly, PolyLine } from "./annotation.js";
+import { Annotation } from "./annotation.js";
 import { CameraParameters } from "./camera_parameters.js";
 import { UiTools } from "./ui_tools.js";
-
-import { FitsImage } from "./layers/fits_image.js";
-import { FitsImageJs } from "./layers/fits_image_js.js";
 
 import { RenderTriangle } from "./render_triangle.js";
 import { Tile } from "./tile.js";
@@ -64,8 +60,6 @@ import { VideoOutputType } from "./video_output_type.js";
 import { ViewMoverSlew, ViewMoverKenBurnsStyle } from "./view_mover.js";
 import { Dialog } from "./utilities/dialog.js";
 
-import { ImageSetLayer } from "./layers/imageset_layer.js";
-
 import { TourPlayer } from "./tours/tour_player.js";
 
 import { LayerManager } from "./layers/layer_manager.js";
@@ -76,563 +70,8 @@ import { TourEditTab } from "./tours/tour_edit.js";
 import { Folder } from "./folder.js";
 import { Planets3d } from "./planets_3d.js";
 import { RenderContext } from "./render_context.js";
+import { ScriptInterface } from "./script_interface.js";
 
-
-// wwtlib.ScriptInterface
-
-export function ScriptInterface() {
-  this._missedReady = false;
-  this.hideTourFeedback = false;
-  this._smoothAnimation = false;
-  this._showCaptions = true;
-}
-
-ScriptInterface._containsFitsLikeExtentsion = function (url) {
-  var lowerCaseUrl = url.toLowerCase();
-  return (ss.endsWith(lowerCaseUrl, 'fits') || ss.endsWith(lowerCaseUrl, 'ftz') || ss.endsWith(lowerCaseUrl, 'fit') || ss.endsWith(lowerCaseUrl, 'fts'));
-};
-
-ScriptInterface._addImageSet = function (name, gotoTarget, loaded, imageset) {
-  if (ss.whitespace(name)) {
-    name = LayerManager.getNextImageSetName();
-  }
-  var imagesetLayer = LayerManager.addImageSetLayerCallback(imageset, name, loaded);
-  if (gotoTarget) {
-    var zoom = imageset._guessZoomSetting(globalRenderContext.viewCamera.zoom);
-    globalWWTControl.gotoRADecZoom(imageset.get_viewCenterX() / 15, imageset.get_viewCenterY(), zoom, false, null);
-  }
-  return imagesetLayer;
-};
-
-ScriptInterface._addFitsLayer = function (url, name, gotoTarget, loaded) {
-  if (ss.whitespace(name)) {
-    name = LayerManager.getNextFitsName();
-  }
-  var imagesetLayer = new ImageSetLayer();
-  var imageset = new Imageset();
-  var wcsLoaded = function (wcsImage) {
-    if ((wcsImage).errored) {
-      return;
-    }
-    var width = ss.truncate(wcsImage.get_sizeX());
-    var height = ss.truncate(wcsImage.get_sizeY());
-    imageset.setInitialParameters(wcsImage.get_description(), wcsImage.get_filename(), 2, 3, 5, Util.getHashCode(wcsImage.get_filename()), 0, 0, wcsImage.get_scaleY(), '.fits', wcsImage.get_scaleX() > 0, '', wcsImage.get_centerX(), wcsImage.get_centerY(), wcsImage.get_rotation(), false, '', false, false, 1, wcsImage.get_referenceX(), wcsImage.get_referenceY(), wcsImage.get_copyright(), wcsImage.get_creditsUrl(), '', '', 0, '');
-    imageset.set_wcsImage(wcsImage);
-    imagesetLayer.set_imageSet(imageset);
-    LayerManager.addFitsImageSetLayer(imagesetLayer, name);
-    if (gotoTarget) {
-      var zoom = imageset._guessZoomSetting(globalRenderContext.viewCamera.zoom);
-      globalWWTControl.gotoRADecZoom(wcsImage.get_viewCenterX() / 15, wcsImage.get_viewCenterY(), zoom, false, null);
-    }
-    if (loaded != null) {
-      loaded(imagesetLayer);
-    }
-  };
-  if (ss.whitespace(name)) {
-    name = LayerManager.getNextFitsName();
-  }
-  if (useGlVersion2) {
-    new FitsImage(imageset, url, null, wcsLoaded);
-  }
-  else {
-    new FitsImageJs(imageset, url, null, wcsLoaded);
-  }
-  return imagesetLayer;
-};
-
-var ScriptInterface$ = {
-  add_ready: function (value) {
-    this.__ready = ss.bindAdd(this.__ready, value);
-  },
-
-  remove_ready: function (value) {
-    this.__ready = ss.bindSub(this.__ready, value);
-  },
-
-  _fireReady: function () {
-    if (this.__ready != null) {
-      this.__ready(this, new ss.EventArgs());
-    } else {
-      this._missedReady = true;
-    }
-  },
-
-  add_collectionLoaded: function (value) {
-    this.__collectionLoaded = ss.bindAdd(this.__collectionLoaded, value);
-  },
-
-  remove_collectionLoaded: function (value) {
-    this.__collectionLoaded = ss.bindSub(this.__collectionLoaded, value);
-  },
-
-  _fireCollectionLoaded: function (url) {
-    if (this.__collectionLoaded != null) {
-      this.__collectionLoaded(this, new CollectionLoadedEventArgs(url));
-    }
-  },
-
-  add_colorPickerDisplay: function (value) {
-    this.__colorPickerDisplay = ss.bindAdd(this.__colorPickerDisplay, value);
-  },
-
-  remove_colorPickerDisplay: function (value) {
-    this.__colorPickerDisplay = ss.bindSub(this.__colorPickerDisplay, value);
-  },
-
-  add_voTableDisplay: function (value) {
-    this.__voTableDisplay = ss.bindAdd(this.__voTableDisplay, value);
-  },
-
-  remove_voTableDisplay: function (value) {
-    this.__voTableDisplay = ss.bindSub(this.__voTableDisplay, value);
-  },
-
-  add_refreshLayerManager: function (value) {
-    this.__refreshLayerManager = ss.bindAdd(this.__refreshLayerManager, value);
-  },
-
-  remove_refreshLayerManager: function (value) {
-    this.__refreshLayerManager = ss.bindSub(this.__refreshLayerManager, value);
-  },
-
-  add_arrived: function (value) {
-    this.__arrived = ss.bindAdd(this.__arrived, value);
-  },
-
-  remove_arrived: function (value) {
-    this.__arrived = ss.bindSub(this.__arrived, value);
-  },
-
-  add_clicked: function (value) {
-    this.__clicked = ss.bindAdd(this.__clicked, value);
-  },
-
-  remove_clicked: function (value) {
-    this.__clicked = ss.bindSub(this.__clicked, value);
-  },
-
-  add_annotationClicked: function (value) {
-    this.__annotationClicked = ss.bindAdd(this.__annotationClicked, value);
-  },
-
-  remove_annotationClicked: function (value) {
-    this.__annotationClicked = ss.bindSub(this.__annotationClicked, value);
-  },
-
-  add_imageryLoaded: function (value) {
-    this.__imageryLoaded = ss.bindAdd(this.__imageryLoaded, value);
-  },
-
-  remove_imageryLoaded: function (value) {
-    this.__imageryLoaded = ss.bindSub(this.__imageryLoaded, value);
-  },
-
-  add_tourReady: function (value) {
-    this.__tourReady = ss.bindAdd(this.__tourReady, value);
-  },
-
-  remove_tourReady: function (value) {
-    this.__tourReady = ss.bindSub(this.__tourReady, value);
-  },
-
-  add_tourError: function (value) {
-    this.__tourError = ss.bindAdd(this.__tourError, value);
-  },
-
-  remove_tourError: function (value) {
-    this.__tourError = ss.bindSub(this.__tourError, value);
-  },
-
-  add_tourPaused: function (value) {
-    this.__tourPaused = ss.bindAdd(this.__tourPaused, value);
-  },
-
-  remove_tourPaused: function (value) {
-    this.__tourPaused = ss.bindSub(this.__tourPaused, value);
-  },
-
-  add_tourResumed: function (value) {
-    this.__tourResumed = ss.bindAdd(this.__tourResumed, value);
-  },
-
-  remove_tourResumed: function (value) {
-    this.__tourResumed = ss.bindSub(this.__tourResumed, value);
-  },
-
-  add_tourEnded: function (value) {
-    this.__tourEnded = ss.bindAdd(this.__tourEnded, value);
-  },
-
-  remove_tourEnded: function (value) {
-    this.__tourEnded = ss.bindSub(this.__tourEnded, value);
-  },
-
-  add_slideChanged: function (value) {
-    this.__slideChanged = ss.bindAdd(this.__slideChanged, value);
-  },
-
-  remove_slideChanged: function (value) {
-    this.__slideChanged = ss.bindSub(this.__slideChanged, value);
-  },
-
-  add_timeScrubberHook: function (value) {
-    this.__timeScrubberHook = ss.bindAdd(this.__timeScrubberHook, value);
-  },
-
-  remove_timeScrubberHook: function (value) {
-    this.__timeScrubberHook = ss.bindSub(this.__timeScrubberHook, value);
-  },
-
-  setTimeScrubberPosition: function (posLeft) {
-    LayerManager.setTimeSliderValue(posLeft);
-  },
-
-  setTimeSlider: function (name, value) {
-    this.__timeScrubberHook(name, value);
-  },
-
-  showColorPicker: function (pickerInstance, e) {
-    if (this.__colorPickerDisplay != null) {
-      this.__colorPickerDisplay(pickerInstance, e);
-    }
-  },
-
-  displayVoTableLayer: function (layer) {
-    if (this.__voTableDisplay != null) {
-      this.__voTableDisplay(layer, new ss.EventArgs());
-    }
-  },
-
-  refreshLayerManagerNow: function () {
-    if (this.__refreshLayerManager != null) {
-      this.__refreshLayerManager(null, new ss.EventArgs());
-    }
-  },
-
-  _fireTourReady: function () {
-    if (this.__tourReady != null) {
-      this.__tourReady(this, new ss.EventArgs());
-    }
-  },
-
-  _fireTourError: function (ex) {
-    if (this.__tourError != null) {
-      this.__tourError(ex, new ss.EventArgs());
-    }
-  },
-
-  _fireTourPaused: function () {
-    if (this.__tourPaused != null) {
-      this.__tourPaused(this, new ss.EventArgs());
-    }
-  },
-
-  _fireTourResume: function () {
-    if (this.__tourResumed != null) {
-      this.__tourResumed(this, new ss.EventArgs());
-    }
-  },
-
-  _fireTourEnded: function () {
-    if (this.__tourEnded != null) {
-      this.__tourEnded(this, new ss.EventArgs());
-    }
-  },
-
-  _fireImageryLoaded: function () {
-    if (this.__imageryLoaded != null) {
-      this.__imageryLoaded(this, new ss.EventArgs());
-    }
-  },
-
-  _fireClick: function (ra, dec) {
-    if (this.__clicked != null) {
-      this.__clicked(this, new ArrivedEventArgs(ra, dec, globalRenderContext.viewCamera.zoom));
-    }
-  },
-
-  _fireArrived: function (ra, dec, zoom) {
-    if (this.__arrived != null) {
-      this.__arrived(this, new ArrivedEventArgs(ra, dec, zoom));
-    }
-  },
-
-  _fireAnnotationclicked: function (RA, Dec, id) {
-    try {
-      if (this.__annotationClicked != null) {
-        this.__annotationClicked(this, new AnnotationClickEventArgs(RA, Dec, id));
-      }
-    }
-    catch ($e1) {
-    }
-  },
-
-  _fireSlideChanged: function (caption) {
-    try {
-      if (this.__slideChanged != null) {
-        this.__slideChanged(this, new SlideChangedEventArgs(caption));
-      }
-    }
-    catch ($e1) {
-    }
-  },
-
-  endInit: function () {
-    if (this._missedReady) {
-      this._fireReady();
-    }
-  },
-
-  gotoRaDecZoom: function (ra, dec, zoom, instant, roll) {
-    if (globalWWTControl != null) {
-      globalWWTControl.gotoRADecZoom(ra / 15, dec, zoom * 6, instant, roll);
-    }
-  },
-
-  setBackgroundImageByName: function (name) {
-    if (globalWWTControl != null) {
-      globalWWTControl.setBackgroundImageByName(name);
-    }
-  },
-
-  addVoTableLayer: function (table) {
-    return LayerManager.addVoTableLayer(table, 'Vo Table');
-  },
-
-  getLayers: function () {
-    return LayerManager.get_layerList();
-  },
-
-  setForegroundImageByName: function (name) {
-    if (globalWWTControl != null) {
-      globalWWTControl.setForegroundImageByName(name);
-      globalRenderContext.viewCamera.opacity = 100;
-    }
-  },
-
-  setForegroundOpacity: function (opacity) {
-    if (globalWWTControl != null) {
-      globalRenderContext.viewCamera.opacity = opacity;
-    }
-  },
-
-  addCatalogHipsByName: function (name) {
-    if (globalWWTControl != null) {
-      globalWWTControl.addCatalogHipsByName(name);
-    }
-  },
-
-  addCatalogHipsByNameWithCallback: function (name, onLoad) {
-    if (globalWWTControl != null) {
-      globalWWTControl.addCatalogHipsByNameWithCallback(name, onLoad);
-    }
-  },
-
-  removeCatalogHipsByName: function (name) {
-    if (globalWWTControl != null) {
-      globalWWTControl.removeCatalogHipsByName(name);
-    }
-  },
-
-  getCatalogHipsDataInView: function (name, limit, onComplete) {
-    if (globalWWTControl != null) {
-      globalWWTControl.getCatalogHipsDataInView(name, limit, onComplete);
-    }
-  },
-
-  setCutsForFits: function (imagesetName, min, max) {
-    if (globalWWTControl != null) {
-      globalWWTControl.setCutsForFits(imagesetName, min, max);
-    }
-  },
-
-  setColorMapForFits: function (imagesetName, colorMapName) {
-    if (globalWWTControl != null) {
-      globalWWTControl.setColorMapForFits(imagesetName, colorMapName);
-    }
-  },
-
-  setScaleTypeForFits: function (imagesetName, scaleType) {
-    if (globalWWTControl != null) {
-      globalWWTControl.setScaleTypeForFits(imagesetName, scaleType);
-    }
-  },
-
-  hideUI: function (hide) { },
-
-  loadTour: function (url) {
-    if (globalWWTControl != null) {
-      globalWWTControl.playTour(url);
-    }
-  },
-
-  loadFits: function (url) {
-    return this.loadFitsLayer(url, '', true, null);
-  },
-
-  loadFitsLayer: function (url, name, gotoTarget, loaded) {
-    return this.addImageSetLayer(url, 'fits', name, gotoTarget, loaded);
-  },
-
-  addImageSetLayer: function (url, mode, name, gotoTarget, loaded) {
-    if (mode != null && mode.toLowerCase() === 'fits') {
-      return ScriptInterface._addFitsLayer(url, name, gotoTarget, loaded);
-    } else if (mode != null && mode.toLowerCase() === 'preloaded') {
-      var imageset = globalWWTControl.getImageSetByUrl(url);
-      if (imageset != null) {
-        return ScriptInterface._addImageSet(name, gotoTarget, loaded, imageset);
-      }
-    } else {
-      var imageset = globalWWTControl.getImageSetByUrl(url);
-      if (imageset != null) {
-        return ScriptInterface._addImageSet(name, gotoTarget, loaded, imageset);
-      }
-      else if (ScriptInterface._containsFitsLikeExtentsion(url)) {
-        return ScriptInterface._addFitsLayer(url, name, gotoTarget, loaded);
-      }
-    }
-    return null;
-  },
-
-  setImageSetLayerOrder: function (id, order) {
-    var layer = LayerManager.get_layerList()[id];
-    if (ss.canCast(layer, ImageSetLayer) && order >= 0) {
-      ss.remove(layerManagerGetAllMaps()[layer.get_referenceFrame()].layers, layer);
-      layerManagerGetAllMaps()[layer.get_referenceFrame()].layers.splice(order, 0, layer);
-    }
-  },
-
-  isUsingWebGl2: function () {
-    return useGlVersion2;
-  },
-
-  get_hideTourFeedback: function () {
-    return this.hideTourFeedback;
-  },
-
-  set_hideTourFeedback: function (value) {
-    this.hideTourFeedback = value;
-    return value;
-  },
-
-  playTour: function () {
-    if (globalWWTControl != null) {
-      globalWWTControl.playCurrentTour();
-    }
-  },
-
-  stopTour: function () {
-    if (globalWWTControl != null) {
-      globalWWTControl.stopCurrentTour();
-    }
-  },
-
-  loadImageCollection: function (url, loadChildFolders) {
-    var $this = this;
-
-    this._imageUrl = url;
-    Wtml.getWtmlFile(url, function () {
-      $this._fireCollectionLoaded(url);
-    }, loadChildFolders);
-  },
-
-  _imageFileLoaded: function () {
-    this._fireCollectionLoaded(this._imageUrl);
-  },
-
-  zoom: function (factor) {
-    if (globalWWTControl != null) {
-      globalWWTControl.zoom(factor);
-    }
-    return;
-  },
-
-  getRA: function () {
-    if (globalWWTControl != null) {
-      return globalRenderContext.get_RA();
-    }
-    return 0;
-  },
-
-  getDec: function () {
-    if (globalWWTControl != null) {
-      return globalRenderContext.get_dec();
-    }
-    return 0;
-  },
-
-  createFolder: function () {
-    var folder = new Folder();
-    return folder;
-  },
-
-  createPolygon: function (fill) {
-    var p = new Poly();
-    p.set_fill(fill);
-    return p;
-  },
-
-  createPolyLine: function (fill) {
-    return new PolyLine();
-  },
-
-  createCircle: function (fill) {
-    var c = new Circle();
-    c.set_fill(fill);
-    return c;
-  },
-
-  addAnnotation: function (annotation) {
-    if (annotation != null && ss.canCast(annotation, Annotation)) {
-      if (globalWWTControl != null) {
-        globalWWTControl._addAnnotation(annotation);
-      }
-    }
-  },
-
-  removeAnnotation: function (annotation) {
-    if (annotation != null) {
-      if (globalWWTControl != null) {
-        globalWWTControl._removeAnnotation(annotation);
-      }
-    }
-  },
-
-  clearAnnotations: function () {
-    if (globalWWTControl != null) {
-      globalWWTControl._clearAnnotations();
-    }
-  },
-
-  get_smoothAnimation: function () {
-    return this._smoothAnimation;
-  },
-
-  set_smoothAnimation: function (value) {
-    this._smoothAnimation = value;
-    return value;
-  },
-
-  get_showCaptions: function () {
-    return this._showCaptions;
-  },
-
-  set_showCaptions: function (value) {
-    this._showCaptions = value;
-    return value;
-  },
-
-  loadVOTable: function (url, useCurrentView) { },
-
-  get_fov: function () {
-    if (globalWWTControl != null) {
-      return globalRenderContext.viewCamera.zoom / 6;
-    }
-    return 60;
-  }
-};
-
-registerType("ScriptInterface", [ScriptInterface, ScriptInterface$, null]);
 
 // wwtlib.FolderDownloadAction
 
@@ -678,6 +117,8 @@ Wtml.getWtmlFile = function (url, complete, loadChildFolders) {
   folderDownloadAction.startingNewFolderLoad(folder);
   return folder;
 };
+
+set_loadWtmlFile(Wtml.getWtmlFile);
 
 Wtml.loadImagesets = function (folder, folderDownloadAction) {
   var children = folder.get_children();
@@ -1996,7 +1437,7 @@ var WWTControl$ = {
         $this._setupComplete();
       }, 0);
     } else {
-      Wtml.getWtmlFile(URLHelpers.singleton.coreDynamicUrl('wwtweb/catalog.aspx?X=ImageSets6'), ss.bind('_setupComplete', this), true);
+      loadWtmlFile(URLHelpers.singleton.coreDynamicUrl('wwtweb/catalog.aspx?X=ImageSets6'), ss.bind('_setupComplete', this), true);
     }
   },
 
@@ -2753,131 +2194,6 @@ var DataVizWizard$ = {
 };
 
 registerType("DataVizWizard", [DataVizWizard, DataVizWizard$, Dialog]);
-
-// wwtlib.SlideChangedEventArgs
-
-export function SlideChangedEventArgs(caption) {
-  ss.EventArgs.call(this);
-  this.set_caption(caption);
-}
-
-var SlideChangedEventArgs$ = {
-  get_caption: function () {
-    return this._caption$2;
-  },
-
-  set_caption: function (value) {
-    this._caption$2 = value;
-    return value;
-  }
-};
-
-registerType("SlideChangedEventArgs", [SlideChangedEventArgs, SlideChangedEventArgs$, ss.EventArgs]);
-
-// wwtlib.ArrivedEventArgs
-
-export function ArrivedEventArgs(ra, dec, zoom) {
-  this._ra$2 = 0;
-  this._dec$2 = 0;
-  this._zoom$2 = 0;
-  ss.EventArgs.call(this);
-  this.set_RA(ra * 15);
-  this.set_dec(dec);
-  this.set_zoom(zoom / 6);
-}
-
-var ArrivedEventArgs$ = {
-  get_RA: function () {
-    return this._ra$2;
-  },
-
-  set_RA: function (value) {
-    this._ra$2 = value;
-    return value;
-  },
-
-  get_dec: function () {
-    return this._dec$2;
-  },
-
-  set_dec: function (value) {
-    this._dec$2 = value;
-    return value;
-  },
-
-  get_zoom: function () {
-    return this._zoom$2;
-  },
-
-  set_zoom: function (value) {
-    this._zoom$2 = value;
-    return value;
-  }
-};
-
-registerType("ArrivedEventArgs", [ArrivedEventArgs, ArrivedEventArgs$, ss.EventArgs]);
-
-// wwtlib.AnnotationClickEventArgs
-
-export function AnnotationClickEventArgs(ra, dec, id) {
-  this._ra$2 = 0;
-  this._dec$2 = 0;
-  ss.EventArgs.call(this);
-  this.set_RA(ra * 15);
-  this.set_dec(dec);
-  this.set_id(id);
-}
-
-var AnnotationClickEventArgs$ = {
-  get_RA: function () {
-    return this._ra$2;
-  },
-
-  set_RA: function (value) {
-    this._ra$2 = value;
-    return value;
-  },
-
-  get_dec: function () {
-    return this._dec$2;
-  },
-
-  set_dec: function (value) {
-    this._dec$2 = value;
-    return value;
-  },
-
-  get_id: function () {
-    return this._id$2;
-  },
-
-  set_id: function (value) {
-    this._id$2 = value;
-    return value;
-  }
-};
-
-registerType("AnnotationClickEventArgs", [AnnotationClickEventArgs, AnnotationClickEventArgs$, ss.EventArgs]);
-
-// wwtlib.CollectionLoadedEventArgs
-
-export function CollectionLoadedEventArgs(url) {
-  ss.EventArgs.call(this);
-  this._url$2 = url;
-}
-
-var CollectionLoadedEventArgs$ = {
-  get_url: function () {
-    return this._url$2;
-  },
-
-  set_url: function (value) {
-    this._url$2 = value;
-    return value;
-  }
-};
-
-registerType("CollectionLoadedEventArgs", [CollectionLoadedEventArgs, CollectionLoadedEventArgs$, ss.EventArgs]);
 
 // Statics with nontrivial initialization:
